Extract shared mesh LLM protocol crate
