+++ .github/workflows/ci.yml
@@ -65,6 +65,7 @@ jobs:
               - 'crates/mesh-api-ffi/**'
               - 'crates/mesh-client/**'
               - 'crates/mesh-llm-identity/**'
+              - 'crates/mesh-llm-protocol/**'
               - 'crates/mesh-llm-types/**'
               - 'sdk/**'
               - 'Package.swift'
