+++ .github/workflows/docker-precheck.yml
@@ -64,6 +64,7 @@ jobs:
           for path in \
             crates/mesh-llm/ \
             crates/mesh-llm-identity/ \
+            crates/mesh-llm-protocol/ \
             crates/mesh-llm-types/ \
             crates/mesh-llm-plugin/ \
             crates/mesh-client/ \
