+++ README.md
@@ -640,6 +640,7 @@ You can also try the hosted demo:
 - [docs/design/VIRTUAL_LLM.md](docs/design/VIRTUAL_LLM.md) for inter-model collaboration design
 - [crates/mesh-llm/README.md](crates/mesh-llm/README.md) for Rust crate structure
 - [crates/mesh-llm-identity/README.md](crates/mesh-llm-identity/README.md) for shared owner identity and envelope crypto
+- [crates/mesh-llm-protocol/README.md](crates/mesh-llm-protocol/README.md) for shared mesh wire protocol types and frame helpers
 - [crates/mesh-llm-types/README.md](crates/mesh-llm-types/README.md) for shared protocol-facing model and mesh data types
 - [crates/mesh-llm-ui/README.md](crates/mesh-llm-ui/README.md) for web console source, build, and embedded asset ownership
 - [ROADMAP.md](ROADMAP.md) for future work
