+++ crates/mesh-client/README.md
@@ -28,6 +28,9 @@ Shared owner identity and envelope crypto are owned by
 `crates/mesh-llm-identity/` and re-exported here for compatibility with existing
 client call sites.
 
+Shared protobuf types and frame helpers are owned by `crates/mesh-llm-protocol/`
+and re-exported here for compatibility with existing client call sites.
+
 Client requests should preserve the full model ref chosen by the caller. Model
 resolution, stage topology, and runtime lifecycle remain server-side mesh
 responsibilities.
