+++ crates/mesh-llm-protocol/src/lib.rs
@@ -1,0 +1,6 @@
+#![forbid(unsafe_code)]
+
+pub mod proto;
+pub mod protocol;
+
+pub use protocol::*;
