+++ crates/mesh-llm-protocol/src/proto/mod.rs
@@ -1,0 +1,1 @@
+pub mod node;
