+++ crates/mesh-llm-protocol/src/protocol/convert.rs
