+++ crates/mesh-llm-protocol/src/protocol/v0.rs
