+++ crates/mesh-llm/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 ansi-to-tui = "8"
 mesh-llm-plugin = { path = "../mesh-llm-plugin" }
 mesh-llm-identity = { path = "../mesh-llm-identity" }
+mesh-llm-protocol = { path = "../mesh-llm-protocol" }
 mesh-llm-types = { path = "../mesh-llm-types" }
 mesh-llm-ui = { path = "../mesh-llm-ui" }
 mesh-client = { package = "mesh-llm-client", path = "../mesh-client", features = ["host-io"] }
@@ -71,7 +72,3 @@ axum = "0.8"
 serial_test = "3"
 mesh-client = { package = "mesh-llm-client", path = "../mesh-client" }
 tempfile = "3"
-
-[build-dependencies]
-prost-build = "0.14"
-protoc-bin-vendored = "3"
