+++ scripts/release-version.sh
@@ -98,7 +98,11 @@ done < <(
     cd "$REPO_ROOT"
     git ls-files \
         'crates/mesh-llm/Cargo.toml' \
+        'crates/mesh-llm-identity/Cargo.toml' \
         'crates/mesh-llm-plugin/Cargo.toml' \
+        'crates/mesh-llm-protocol/Cargo.toml' \
+        'crates/mesh-llm-types/Cargo.toml' \
+        'crates/mesh-llm-ui/Cargo.toml' \
         'crates/mesh-api/Cargo.toml' \
         'crates/mesh-client/Cargo.toml' \
         | sort -u
